+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog].
 ### Fixed
 - Unable to add elements to Setting Group Group `#119`
 - Unable to cancel a build `#120`
+- Uploader may freezes after changing platform `#123`
 
 ### Security